+++ pymelcloud/__init__.py
@@ -1,8 +1,7 @@
 """MELCloud module"""
 from aiohttp import ClientSession
-import asyncio
 from datetime import timedelta
-from typing import List, Optional, Type
+from typing import Dict, List, Optional
 
 from pymelcloud.client import Client as _Client, login as _login
 from pymelcloud.device import Device
@@ -18,22 +17,22 @@ async def login(
     password: str,
     session: Optional[ClientSession] = None,
     *,
-    conf_update_interval: Optional[timedelta] = None,
-    device_set_debounce: Optional[timedelta] = None,
+    conf_update_interval: Optional[timedelta] = timedelta(minutes=5),
+    device_set_debounce: Optional[timedelta] = timedelta(seconds=1),
 ) -> str:
     """
     Log in to MELCloud with given credentials.
     
     Returns access token.
     """
-    client = await _login(
+    _client = await _login(
         email,
         password,
         session,
-        conf_update_interval=timedelta(minutes=5),
-        device_set_debounce=timedelta(seconds=1),
+        conf_update_interval=conf_update_interval,
+        device_set_debounce=device_set_debounce,
     )
-    return client.token
+    return _client.token
 
 
 async def get_devices(
@@ -42,24 +41,24 @@ async def get_devices(
     *,
     conf_update_interval=timedelta(minutes=5),
     device_set_debounce=timedelta(seconds=1),
-) -> List[Type[Device]]:
+) -> Dict[str, List[Device]]:
     """Initialize Devices available with the token."""
-    client = _Client(
+    _client = _Client(
         token,
         session,
         conf_update_interval=conf_update_interval,
         device_set_debounce=device_set_debounce,
     )
-    await client.update_confs()
+    await _client.update_confs()
     return {
         DEVICE_TYPE_ATA: [
-            AtaDevice(conf, client, set_debounce=device_set_debounce)
-            for conf in client.device_confs
-            if conf.get("Device", {}).get("DeviceType", -1) == 0
+            AtaDevice(conf, _client, set_debounce=device_set_debounce)
+            for conf in _client.device_confs
+            if conf.get("Device", {}).get("DeviceType") == 0
         ],
         DEVICE_TYPE_ATW: [
-            AtwDevice(conf, client, set_debounce=device_set_debounce)
-            for conf in client.device_confs
-            if conf.get("Device", {}).get("DeviceType", -1) == 1
+            AtwDevice(conf, _client, set_debounce=device_set_debounce)
+            for conf in _client.device_confs
+            if conf.get("Device", {}).get("DeviceType") == 1
         ],
     }