+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix `get_devices` type hints.
+- Fix `conf_update_interval` and `device_set_debounce` forwarding in `login`.
 
 ## [2.0.0] - 2020-02-08
 ### Added